requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Werkzeug==3.0.3
openai==1.51.0
groq==0.4.1
python-dotenv==1.0.1
Pillow==10.4.0
